[refactoring] rename MandateData to OutputRow::Mandate
Make a namespace for output rows: initially for the Mandate rows, but
with rows for the inception/abolition dates, and successor/predecessor
rows to follow.

This is a step towards making these more consistent.
