+++ lib/wikidata_position_history/report.rb
@@ -1,47 +1,6 @@
 # frozen_string_literal: true
 
 module WikidataPositionHistory
-  # Date for a single mandate row, to be passed to the report template
-  class MandateData
-    CHECKS = [Check::MissingFields, Check::Overlap,
-              Check::WrongPredecessor, Check::MissingPredecessor,
-              Check::WrongSuccessor, Check::MissingSuccessor].freeze
-
-    def initialize(later, current, earlier)
-      @later = later
-      @current = current
-      @earlier = earlier
-    end
-
-    def ordinal_string
-      ordinal = current.ordinal or return ''
-      "#{ordinal}."
-    end
-
-    def person
-      current.item
-    end
-
-    def dates
-      dates = [current.start_date, current.end_date]
-      return '' if dates.compact.empty?
-
-      dates.join(' – ')
-    end
-
-    def acting?
-      current.acting?
-    end
-
-    def warnings
-      CHECKS.map { |klass| klass.new(later, current, earlier) }.select(&:problem?)
-    end
-
-    private
-
-    attr_reader :later, :current, :earlier
-  end
-
   # Data about the position itself, to be passed to the report template
   class Metadata
     # simplified version of a WikidataPositionHistory::Check
@@ -205,7 +164,7 @@ module WikidataPositionHistory
     def table_rows
       padded_mandates.each_cons(3).map do |later, current, earlier|
         {
-          mandate: MandateData.new(later, current, earlier),
+          mandate: OutputRow::Mandate.new(later, current, earlier),
           bio:     biodata_for(current.officeholder),
         }
       end
